requirements.txt
git+https://github.com/adsabs/ADSMicroserviceUtils.git@v1.1.2
adsutils==1.0.33
flask-discoverer==0.1.4
Flask-RESTful==0.3.6
Flask-Testing==0.6.2
Werkzeug==0.11.15

